Basicly made this work properly
