+++ config.json
@@ -1,8 +1,9 @@
 {
-	"token":"YOUR_BOT_TOKEN_HERE",
+	"token":"Njc4MDczNTcwMTczMjU1Njgz.Xkdiww.mOAa5TcpzIYihCcEOCAVrSKyb6E",
     "commandPrefix":"/",
-    "ip":"YOUR_SERVER_IP_HERE",
-    "port":"25565",
+    "ip":"100.11.56.150",
+    "port":"65535",
     "pingInterval":"30",
     "embedColor":"7289DA"
+    
 }
