+++ start.sh
@@ -1,5 +1,5 @@
 # /bin/sh
-until node mcbot.js; do
+until npm start; do
     echo "Bot stopped with exit code $?. Restarting..." >&2
     sleep 1
 done
